% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the FARS data on State map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state-number of the state plotted on the map (A numeric value).}

\item{year}{the year of FARS data plotted on the map.}
}
\value{
State map with FARS data of a specific year.  If no data to plot, a message and NULL will return.
}
\description{
This function plot FARS data on a state map for a specific year.
}
\examples{
        \dontrun{fars_map_state(47,2013)}

}
