% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS files by a list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years}
}
\value{
a data frame with 2 values: month and year. Return a warning message and NULL if the file does not exist.
}
\description{
This function will take a list of years as input and read the corresponding FARS files of those years.
}
\examples{
    \dontrun{fars_read_years(2013:2015)}

}
