% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of the year for the data file.}
}
\value{
A string of the file name of the input year
}
\description{
This function will make a filename based on the input year, which requires a numerical input.
}
\examples{
    \dontrun{make_filename(2013)}

}
